refactor(bn256): delete bn256 module and move tests to better locations
