+++ powersoftau/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod accumulator;
 pub mod batched_accumulator;
-pub mod bn256;
 pub mod keypair;
 pub mod parameters;
 pub mod utils;
